pyproject.toml
[tool.poetry]
name = "patch-notes"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
langchain = "0.2.1"
langchain-openai = "0.1.7"
python-dotenv = "^1.0.1"
beautifulsoup4 = "4.12.3"
langchain-community = "0.2.1"
chromadb = "^0.5.0"
asyncio = "^3.4.3"
langgraph = "^0.0.55"
langchain-experimental = "^0.0.59"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
